ESC-688 consolidate multiple action builders classes into a single ActionBuilders class
