+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/BaseController.scala
@@ -17,8 +17,8 @@
 package uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 
 import play.api.data.Forms.{mapping, text}
-import play.api.data.{Form, Mapping}
 import play.api.data.validation.{Constraint, Invalid, Valid}
+import play.api.data.{Form, Mapping}
 import play.api.i18n.I18nSupport
 import play.api.mvc._
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.FormValues
